webservices/es_bulk/requirements.txt
elasticsearch>=6.0.0,<7.0.0
pymysql==0.7.9
pendulum==1.1.1
raven==6.0.0
pyyaml==3.12
click==6.7
prettytable==0.7.2